Implement a stage0 compiler
This adds a "compiler", adapted from lisp-in-small-pieces code. The
compiler takes a file or stdin and translates it to a symbolic
representation of instructions.

The second part of this will be to add support in the Rust code for
translating the symbolic representation into its own values; and the
third part will be to get both of these to the point where the stage0
compiler can be compiled itself and included in the Rust code,
becoming a stage1.