Compile top-level definitions
This adds a new entry point to the pretreatment `meaning-toplevel`,
which understands top-level definitions (though not those in `(begin
...)`, yet).

This needs a bit of monkeying around with the top-level
environment. Specifically, it's necessary when compiling to assume
that top-level bindings might be defined later, so a provisional
binding must be supplied when asked, so compilation can continue. This
applies as much to a free variable nested within code somewhere.

The `compile` entry point now compiles all the forms it can read,
whether from stdin or a file, and makes a program out of the whole
lot.