Finish translation and implement more opcodes
This completes the translation from "intermediate" symbolic form, to
bytecode to be consumed by the interpreter.

It also implements a few more of the opcodes. To accommodate set!,
frames need to use RefCells for arguments.

There's no reason to supply storage for globals to the VM, at least at
present, so I've removed that contructor argument. (Predefined values
on the other hand will need to be provided, soon.)