Implement last function machinery, and primitives
This adds (among others) allocate-dotted-frame and pop-cons-argument!,
which are used in the niche situation of a closed, dotted application,
like

    ((lambda vargs (apply + vargs)) 1 2 3)

(which gets compiled to: push all the args, allocate a frame with nil
in the last slot, pop-cons all the values into that slot, put the
frame in the env register then execute the lambda body).

This also adds in all the primitives that have their own opcodes
(list, apply, and call/cc (if I do it) will be predefined but don't
get their own opcodes). Oh and no `read` yet, because the parser
operates on strings rather than streams, at present.