Implement `display`
This adds a native implementation for `display`; this could be
implemented in the language itself, but it's convenient for now to
have it as a primitive.

I've added an output_stream field int he VM to help with testing; a
fuller implementation would use an output port which could be
replaced, but I don't have ports yet.

For apply and possibly list, I'll need access to the VM registers, so
this adds another kind of primitive which just has access to the
registers. (I could probably rewrite the other kinds of primitive in
terms of this one, now I think about it).