Use just one variant for native procedures
Ultimately, native procedures do stuff with the registers, so it's
possible to cut down the number of structs and Value:: variants for
representing them to just one.

This comes at the cost of defining two functions for most primitives
-- one with ValueRef arguments and return value, to be called
directly, and one with a VM argument to be made into a Value::Native.