Fix over-eager frame allocation
Early on, I deviated slightly from the scheme given in LiSP and
assigned frames their next pointer when allocating them. This was a
mistake, because the frame is allocated before entering a closure (so
the arguments can be popped into it), but the environment is extended
_after_ the closure's environment is restored.

This meant that references to closed-over variables got the wrong
value, since they came from the environment of the caller, rather than
the environment of the closure.