Expand special forms used by compiler, in compiler
To be able to compile itself, the compiler needs to know about the
special forms `let`, `let*`, `and`, and so on. To support those, the
`meaning` procedure now looks closely at any parentheticals to see if
they are listed as a special form, and uses th given expander if so.

For the expanders, it's handy to use `quasiquote`; so, that gets
implemented too. And, to support `let loop`, may as well have `letrec`
even though it's not directly used in the compiler code.

There's also a bunch of procedures that are used by the compiler
(including in the expanders) that are available in the bootstrapping
system (e.g., SISC), but are not primitives. These go in builtin.ss,
so they can be compiled and included separately when it comes to
building the compiler into the runtime.