Compile internal definitions as a letrec
This adds support to the compiler for definitions internal to a
closure body, e.g., the definition of `less-than-ten` in the
following:

    (define (foo n)
      (define (less-than-10 i) (< i 10))
      (if (less-than-ten n)
        (display "less than 10")
        (display "greater than or equal to 10")))

This is equivalent to a letrec (which the expander knows how to do),
but because the definitions can have different shapes, it is not
straight-forward to create an expression to be expanded. Instead,
having done the work of getting all the definitions into the right
form, I compile it directly.